+++ src/popup/components/Main/styles/FeeInfo.module.scss
