+++ src/popup/components/Main/styles/LastBlockInfo.module.scss
