+++ tsconfig.json
@@ -15,7 +15,7 @@
         "rootDir": "src",
         "outDir": "dist/js",
         "noEmitOnError": true,
-        "jsx": "react",
+        "jsx": "react-jsx",
         "typeRoots": [ "node_modules/@types", "./custom_typing"]
     },
     "include": [
